requirements.txt
# Pionex Trading Bot GUI - Requirements (Compatible Versions)
# Install with: pip install -r requirements.txt

# Build tools (Required for compatibility)
setuptools>=65.0.0
wheel>=0.38.0

# Core Flask Framework (Compatible versions)
Flask>=2.2.0,<2.4.0
Flask-SocketIO>=5.2.0,<5.4.0
python-socketio>=5.7.0,<5.9.0
python-engineio>=4.6.0,<4.8.0
Werkzeug>=2.2.0,<2.4.0

# Environment and Configuration
python-dotenv>=0.19.0,<1.1.0
PyYAML>=6.0.0,<7.0.0

# Database
SQLAlchemy>=2.0.0,<2.1.0

# HTTP Requests
requests>=2.28.0,<2.32.0
urllib3>=1.26.0,<2.1.0

# Cryptography and Security
cryptography>=40.0.0,<42.0.0

# Data Processing (Compatible versions)
pandas>=1.5.0,<2.2.0
numpy>=1.24.0,<1.27.0

# Technical Analysis
ta>=0.10.0,<0.11.0
scipy>=1.10.0,<1.12.0

# WebSocket Support
websocket-client>=1.5.0,<1.7.0

# Additional Utilities
schedule>=1.2.0,<1.3.0
psutil>=5.9.0,<5.10.0

# Production Server
gunicorn>=20.1.0,<21.3.0
eventlet>=0.33.0,<0.34.0 